chore: add hydration error handling and better system for desktop and mobile versions
