+++ packages/react-app-revamp/helpers/isTokenShorthand.ts
@@ -1,0 +1,8 @@
+import { TokenConfig } from "./tokens";
+
+export const isTokenShorthand = (allTokens: TokenConfig[], value: string) => {
+  const tokenShorthands = allTokens.map((token: { symbol: string }) => token.symbol);
+  const normalizedValue = value.replace("$", ""); // Remove $ sign from the value
+
+  return tokenShorthands?.includes(normalizedValue);
+};
