+++ packages/react-app-revamp/helpers/mediaQuery.ts
@@ -1,0 +1,5 @@
+import dynamic from "next/dynamic";
+
+export const MediaQuery = dynamic(() => import("react-responsive"), {
+  ssr: false,
+});
