+++ packages/react-app-revamp/package.json
@@ -55,6 +55,7 @@
     "react-html-props": "^1.0.33",
     "react-loading-skeleton": "^3.2.0",
     "react-markdown": "^8.0.7",
+    "react-responsive": "^9.0.2",
     "react-spinners": "^0.13.8",
     "react-toastify": "^9.1.3",
     "react-tooltip": "^5.13.0",
